Update dorahacks-quadratic-funding.md (#370)
We completed frontend integration with HackerLink. However due to tight schedules we didn't create a demo page based on the substrate-frontend-template (task #2 of Milestone-2). Another reason why we didn't create the simple page is that HackerLink has a complete infrastructure for user registration, project upload / display, etc. The HackerLink code base is not open source. We would like to provide a complete demo of how quadratic funding grant is functioning on a parachain or potentially on Polkadot / Kusama relaychains in the future. Therefore we would like to still submit a milestone. Given that our frontend is not open source and we are only providing a demonstration site that uses our milestone-1 code to provide a service for quadratic funding grant, we are reducing the price to 0 accordingly.