update the delivery (#1293)
* update the delivery

I have deleted requirements 3 and requirement 4 for the following reasons: 
The main reason for changing the delivery is that we have over-promised for milestone 2 and underestimated the engineering cost.  We have worked on the UI optimization for a couple of months, and the target is to surpass Duen Analytics in user experience.  Web3Go currently focuses on data visualization and analysis so that is the reason for deleting the labeling system and GraphQL-based query.
It does not mean that we will not implement it. instead, it is very important for a data analysis project, but it will take a long time, I would say a couple of years. 
So I would suggest for milestone 2,  let's focus on delivering a user-friendly dashboard-creating tool. 
The estimated delivery date is by the end of this year. 
And here is the preview version which is under development. we have integrated all Moonbeam data into it. https://dev-v2.web3go.xyz/layout/home

* Update applications/Web3Go.md

it is ok for us! thanks.

Co-authored-by: Sebastian Müller <sebastian@web3.foundation>

Co-authored-by: Sebastian Müller <sebastian@web3.foundation>